create account folder and move components
