+++ app/src/components/account/RevealSeedModal.vue
