+++ extension/src/components/SessionAccounts.vue
@@ -25,7 +25,7 @@
 </template>
 
 <script>
-import AccountList from 'common/AccountList'
+import AccountList from 'account/AccountList'
 import SessionFrame from 'common/SessionFrame'
 import config from 'config'
