+++ extension/src/popup/router/routes.js
@@ -51,7 +51,7 @@ export default (store) => {
     {
       path: `/reveal/:address`,
       name: `reveal`,
-      component: require('common/RevealSeedModal').default
+      component: require('account/RevealSeedModal').default
     },
     {
       path: `/select-network`,
